@@ -18,17 +18,16 @@
   },
   "private": false,
   "dependencies": {
-    "@angular/animations": "^7.2.9",
+    "@angular/animations": "^8.1.0-next.3",
     "@angular/cdk": "^7.2.1",
-    "@angular/common": "~7.2.9",
-    "@angular/compiler": "~7.2.9",
-    "@angular/core": "~7.2.9",
-    "@angular/forms": "~7.2.9",
-    "@angular/http": "~7.2.9",
+    "@angular/common": "~8.1.0-next.3",
+    "@angular/compiler": "~8.1.0-next.3",
+    "@angular/core": "~8.0.2",
+    "@angular/forms": "~8.1.0-next.3",
     "@angular/material": "^7.2.0",
-    "@angular/platform-browser": "~7.2.9",
-    "@angular/platform-browser-dynamic": "~7.2.9",
-    "@angular/router": "~7.2.9",
+    "@angular/platform-browser": "~8.1.0-next.3",
+    "@angular/platform-browser-dynamic": "~8.1.0-next.3",
+    "@angular/router": "~8.1.0-next.3",
     "@types/quill": "^1.3.6",
     "ag-grid-angular": "^20.0.0",
     "ag-grid-community": "^20.0.0",
@@ -39,16 +38,16 @@
     "lodash": "^4.17.11",
     "ng-mocks": "^7.5.0",
     "quill": "^1.3.6",
-    "rxjs": "~6.4.0",
-    "zone.js": "~0.8.29"
+    "rxjs": "~6.5.2",
+    "zone.js": "~0.9.1"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.12.1",
-    "@angular-devkit/build-ng-packagr": "~0.12.1",
+    "@angular-devkit/build-angular": "~0.800.0",
+    "@angular-devkit/build-ng-packagr": "~0.800.0",
     "@angular-devkit/core": "^7.3.6",
-    "@angular/cli": "~7.2.1",
-    "@angular/compiler-cli": "~7.2.9",
-    "@angular/language-service": "~7.2.9",
+    "@angular/cli": "~8.0.3",
+    "@angular/compiler-cli": "~8.1.0-next.3",
+    "@angular/language-service": "~8.1.0-next.3",
     "@babel/core": "^7.2.2",
     "@storybook/addon-actions": "^5.0.10",
     "@storybook/addon-knobs": "^5.0.10",
@@ -61,7 +60,7 @@
     "@types/jasminewd2": "~2.0.3",
     "@types/node": "^8.9.5",
     "babel-loader": "^8.0.5",
-    "codelyzer": "~4.5.0",
+    "codelyzer": "^5.0.1",
     "cpx": "^1.5.0",
     "jasmine-core": "~2.99.1",
     "jasmine-marbles": "^0.4.1",
@@ -71,14 +70,14 @@
     "karma-coverage-istanbul-reporter": "~2.0.1",
     "karma-jasmine": "~1.1.2",
     "karma-jasmine-html-reporter": "^0.2.2",
-    "ng-packagr": "^4.7.1",
+    "ng-packagr": "^5.1.0",
     "protractor": "~5.4.1",
     "ts-node": "~7.0.1",
     "tsconfig-paths": "^3.7.0",
-    "tsickle": ">=0.29.0",
-    "tslib": "^1.9.0",
+    "tsickle": "^0.35.0",
+    "tslib": "^1.10.0",
     "tslint": "~5.12.1",
-    "typescript": "~3.2.4",
+    "typescript": "~3.4.5",
     "prettier": "^1.16.4"
   },
   "publishConfig": {
