@@ -31,9 +31,9 @@ export class InputComponent extends BaseInputElement implements AfterViewInit {
   constructor(private DOM: DOMhelpers) {
     super();
   }
-  @ViewChild('input') input: ElementRef;
-  @ViewChild('prefix') prefix: ElementRef;
-  @ViewChild('suffix') suffix: ElementRef;
+  @ViewChild('input', { static: true }) input: ElementRef;
+  @ViewChild('prefix', { static: false }) prefix: ElementRef;
+  @ViewChild('suffix', { static: false }) suffix: ElementRef;
 
   @Input() hasPrefix = false;
   public showPrefix = true;
