@@ -51,10 +51,10 @@ export abstract class RTEformElement extends BaseFormElement
   @Input() private formControlName: any;
   @Input() private formControl: any;
 
-  @ViewChild('quillEditor') protected quillEditor: ElementRef;
-  @ViewChild('toolbar') protected toolbar: ElementRef;
-  @ViewChild('suffix') protected suffix: ElementRef;
-  @ViewChild('sizePanel') protected sizePanel: PanelComponent;
+  @ViewChild('quillEditor', { static: true }) protected quillEditor: ElementRef;
+  @ViewChild('toolbar', { static: true }) protected toolbar: ElementRef;
+  @ViewChild('suffix', { static: false }) protected suffix: ElementRef;
+  @ViewChild('sizePanel', { static: false }) protected sizePanel: PanelComponent;
 
   @Output() blurred: EventEmitter<any> = new EventEmitter<any>();
   @Output() focused: EventEmitter<any> = new EventEmitter<any>();
